+++ test/webapi/ows/stac/test_controllers.py
@@ -22,28 +22,57 @@
 
 import unittest
 
+from xcube.webapi.ows.stac.config import DEFAULT_CATALOG_DESCRIPTION
+from xcube.webapi.ows.stac.config import DEFAULT_CATALOG_ID
+from xcube.webapi.ows.stac.config import DEFAULT_CATALOG_TITLE
+from xcube.webapi.ows.stac.config import DEFAULT_COLLECTION_DESCRIPTION
+from xcube.webapi.ows.stac.config import DEFAULT_COLLECTION_ID
+from xcube.webapi.ows.stac.config import DEFAULT_COLLECTION_TITLE
+from xcube.webapi.ows.stac.controllers import STAC_VERSION
 from xcube.webapi.ows.stac.controllers import get_collection
 from xcube.webapi.ows.stac.controllers import get_collection_item
 from xcube.webapi.ows.stac.controllers import get_collection_items
 from xcube.webapi.ows.stac.controllers import get_collections
 from xcube.webapi.ows.stac.controllers import get_conformance
 from xcube.webapi.ows.stac.controllers import get_root
-
 from .test_context import get_stac_ctx
 
 BASE_URL = "http://localhost:8080"
+
+EXPECTED_COLLECTION = {
+    'id': DEFAULT_COLLECTION_ID,
+    'title': DEFAULT_COLLECTION_TITLE,
+    'description': DEFAULT_COLLECTION_DESCRIPTION,
+    'stac_version': STAC_VERSION,
+    'stac_extensions': ['xcube'],
+    'summaries': {},
+    'extent': {},
+    'keywords': [],
+    'license': 'proprietary',
+    'links': [
+        {
+            'href': f'{BASE_URL}/catalog/collections/{DEFAULT_COLLECTION_ID}',
+            'rel': 'self'
+        },
+        {
+            'href': f'{BASE_URL}/catalog/collections',
+            'rel': 'root'
+        }
+    ],
+    'providers': [],
+}
 
 
 class StacControllersTest(unittest.TestCase):
     def test_get_collection_item(self):
         result = get_collection_item(get_stac_ctx().datasets_ctx, BASE_URL,
-                                     "datacubes", "demo-1w")
+                                     DEFAULT_COLLECTION_ID, "demo-1w")
         self.assertIsInstance(result, dict)
         # TODO (forman): add more assertions
 
     def test_get_collection_items(self):
         result = get_collection_items(get_stac_ctx().datasets_ctx, BASE_URL,
-                                      "datacubes")
+                                      DEFAULT_COLLECTION_ID)
         self.assertIsInstance(result, dict)
         self.assertIn('features', result)
 
@@ -54,7 +83,7 @@ class StacControllersTest(unittest.TestCase):
         for feature in features:
             self.assertIsInstance(feature, dict)
             self.assertEqual('Feature', feature.get('type'))
-            self.assertEqual('datacubes', feature.get('collection'))
+            self.assertEqual(DEFAULT_COLLECTION_ID, feature.get('collection'))
             self.assertIsInstance(feature.get('bbox'), list)
             self.assertIsInstance(feature.get('properties'), dict)
             self.assertIsInstance(feature.get('geometry'), dict)
@@ -68,66 +97,12 @@ class StacControllersTest(unittest.TestCase):
 
     def test_get_collection(self):
         result = get_collection(get_stac_ctx().datasets_ctx, BASE_URL,
-                                "datacubes")
-        self.assertEqual(
-            {
-                'id': 'datacubes',
-                'title': 'Data cubes',
-                'description': '',
-                'stac_version': '0.9.0',
-                'stac_extensions': ['xcube'],
-                'summaries': {},
-                'extent': {},
-                'keywords': [],
-                'license': 'proprietary',
-                'links': [
-                    {
-                        'href': f'{BASE_URL}/catalog/collections'
-                                f'/datacubes',
-                        'rel': 'self'
-                    },
-                    {
-                        'href': f'{BASE_URL}/catalog/collections',
-                        'rel': 'root'
-                    }
-                ],
-                'providers': [],
-            },
-            result
-        )
+                                DEFAULT_COLLECTION_ID)
+        self.assertEqual(EXPECTED_COLLECTION, result)
 
     def test_get_collections(self):
         result = get_collections(get_stac_ctx().datasets_ctx, BASE_URL)
-        self.assertEqual(
-            {
-                'collections': [
-                    {
-                        'id': 'datacubes',
-                        'title': 'Data cubes',
-                        'description': '',
-                        'stac_version': '0.9.0',
-                        'stac_extensions': ['xcube'],
-                        'summaries': {},
-                        'extent': {},
-                        'keywords': [],
-                        'license': 'proprietary',
-                        'links': [
-                            {
-                                'href': f'{BASE_URL}/catalog/collections'
-                                        f'/datacubes',
-                                'rel': 'self'
-                            },
-                            {
-                                'href': f'{BASE_URL}/catalog/collections',
-                                'rel': 'root'
-                            }
-                        ],
-                        'providers': [],
-                    }
-                ]
-            },
-            result
-        )
+        self.assertEqual({'collections': [EXPECTED_COLLECTION]}, result)
 
     def test_get_conformance(self):
         result = get_conformance(get_stac_ctx().datasets_ctx)
@@ -148,34 +123,34 @@ class StacControllersTest(unittest.TestCase):
         result = get_root(get_stac_ctx().datasets_ctx, BASE_URL)
         self.assertEqual(
             {
-                'stac_version': '0.9.0', 'id': 'xcube-server',
-                'title': 'xcube Server',
-                'description': 'Catalog of datasets and places'
-                               ' served by xcube.',
+                'stac_version': STAC_VERSION,
+                'id': DEFAULT_CATALOG_ID,
+                'title': DEFAULT_CATALOG_TITLE,
+                'description': DEFAULT_CATALOG_DESCRIPTION,
                 'links': [
                     {'rel': 'self',
-                     'href': 'http://localhost:8080/catalog',
+                     'href': f'{BASE_URL}/catalog',
                      'type': 'application/json',
                      'title': 'this document'},
                     {'rel': 'service-desc',
-                     'href': 'http://localhost:8080/openapi.json',
+                     'href': f'{BASE_URL}/openapi.json',
                      'type': 'application/vnd.oai.openapi+json;version=3.0',
                      'title': 'the API definition'},
                     {'rel': 'service-doc',
-                     'href': 'http://localhost:8080/openapi.html',
+                     'href': f'{BASE_URL}/openapi.html',
                      'type': 'text/html',
                      'title': 'the API documentation'},
                     {'rel': 'conformance',
-                     'href': 'http://localhost:8080/catalog/conformance',
+                     'href': f'{BASE_URL}/catalog/conformance',
                      'type': 'application/json',
                      'title': 'OGC API conformance classes implemented'
                               ' by this server'},
                     {'rel': 'data',
-                     'href': 'http://localhost:8080/catalog/collections',
+                     'href': f'{BASE_URL}/catalog/collections',
                      'type': 'application/json',
                      'title': 'Information about the feature collections'},
                     {'rel': 'search',
-                     'href': 'http://localhost:8080/catalog/search',
+                     'href': f'{BASE_URL}/catalog/search',
                      'type': 'application/json',
                      'title': 'Search across feature collections'}
                 ]