+++ test/webapi/ows/stac/test_routes.py
@@ -23,6 +23,7 @@ from ...helpers import RoutesTestCase
 
 
 class StacRoutesTest(RoutesTestCase):
+    """STAC endpoints smoke tests"""
 
     def test_fetch_catalog(self):
         response = self.fetch('/catalog')
@@ -37,16 +38,24 @@ class StacRoutesTest(RoutesTestCase):
         self.assertResponseOK(response)
 
     def test_fetch_catalog_collection(self):
-        response = self.fetch('/catalog/collections/datacubes')
+        response = self.fetch('/catalog/collections/datasets')
         self.assertResponseOK(response)
+        response = self.fetch('/catalog/collections/datacubes')
+        self.assertResourceNotFoundResponse(response)
 
     def test_fetch_catalog_collection_items(self):
-        response = self.fetch('/catalog/collections/datacubes/items')
+        response = self.fetch('/catalog/collections/datasets/items')
         self.assertResponseOK(response)
+        response = self.fetch('/catalog/collections/datacubes')
+        self.assertResourceNotFoundResponse(response)
 
     def test_fetch_catalog_collection_item(self):
-        response = self.fetch('/catalog/collections/datacubes/items/demo')
+        response = self.fetch('/catalog/collections/datasets/items/demo')
         self.assertResponseOK(response)
+        response = self.fetch('/catalog/collections/datasets/items/demox')
+        self.assertResourceNotFoundResponse(response)
+        response = self.fetch('/catalog/collections/datacubes/items/demo')
+        self.assertResourceNotFoundResponse(response)
 
     def test_fetch_catalog_search_by_kw(self):
         response = self.fetch('/catalog/search', method='GET')